Move Bullet to root components folder
